python_backend/requirements.txt
 Flask==2.0.1
Flask-Cors==3.0.10
mysql-connector-python==8.0.26
python-dotenv==0.19.0
PyJWT==2.1.0
bcrypt==3.2.0